✅ select-game file path.
